requirements.txt
# API de Logística - Dependencias de Producción
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
pydantic==2.11.7
pydantic-settings==2.3.0
python-dotenv==1.1.1

# Solo para desarrollo/testing (comentado para producción)
# pytest==8.4.1
